Rearrenge views folders to SEPARATE students-views from instructors-views
