+++ src/views/Instructors/Feedback.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import MarkdownFeedback from "../components/MarkdownFeedback";
+import MarkdownFeedback from "../../components/MarkdownFeedback";
 // import PropTypes from 'prop-types';
 
 class Feedback extends Component {
